hot fix y refactorizacion de la logica del chat en la WA
